✨ [#83] Dto 세분화 및 수정

+++ backend/src/main/java/com/Devim/backend/domain/comment/Comment.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @Data
 public class Comment {
 	private long commentNo;
-	private int boardNo;
+	private long boardNo;
 	private String commentContent;
 	private int writer;
 	private LocalDateTime createdDt;
